(refactor) Refactor login and remove unused functions
