+++ client/src/actions/action_types.js
@@ -13,10 +13,6 @@ export const REFRESH_ALL_BOARDS = 'REFRESH_ALL_BOARDS';
 
 
 // Auth Actions
-export const LOGIN_REQUEST = 'LOGIN_REQUEST';
-export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
-export const LOGIN_FAILURE = 'LOGIN_FAILURE';
 export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-export const SHOW_LOCK = 'SHOW_LOCK';
 export const LOCK_SUCCESS = 'LOCK_SUCCESS';
 export const LOCK_ERROR = 'LOCK_ERROR';
