+++ client/src/actions/auth_actions.js
@@ -1,64 +1,15 @@
-// import axios from 'axios';
-import { LOGIN_REQUEST, LOGIN_SUCCESS, LOGIN_FAILURE, SHOW_LOCK, LOCK_SUCCESS, LOCK_ERROR, LOGOUT_REQUEST, LOGOUT_SUCCESS } from './action_types';
-
-export function requestLogin(creds) {
-  return {
-    type: LOGIN_REQUEST,
-    isFetching: true,
-    isAuthenticated: false,
-    creds,
-  };
-}
-
-export function receiveLogin(user) {
-  return {
-    type: LOGIN_SUCCESS,
-    isFetching: false,
-    isAuthenticated: true,
-    id_token: user.id_token,
-  };
-}
-
-export function loginError(message) {
-  return {
-    type: LOGIN_FAILURE,
-    isFetching: false,
-    isAuthenticated: false,
-    message,
-  };
-}
-
-export function requestLogout() {
-  return {
-    type: LOGOUT_REQUEST,
-    isFetching: true,
-    isAuthenticated: true,
-  };
-}
+import * as types from './action_types';
 
 export function receiveLogout() {
   return {
-    type: LOGOUT_SUCCESS,
-    isFetching: false,
+    type: types.LOGOUT_SUCCESS,
     isAuthenticated: false,
-  };
-}
-
-export function logoutUser() {
-  const lock = new Auth0Lock('Wxg5EhfSanTioiIfdLOdlCwdKoK7twjn', 'app50347872.auth0.com');
-  lock.logout({ ref: window.location.href });
-  // console.log(window.location.href);
-}
-
-export function showLock() {
-  return {
-    type: SHOW_LOCK,
   };
 }
 
 export function lockSuccess(profile, token) {
   return {
-    type: LOCK_SUCCESS,
+    type: types.LOCK_SUCCESS,
     profile,
     token,
   };
@@ -66,7 +17,7 @@ export function lockSuccess(profile, token) {
 
 export function lockError(err) {
   return {
-    type: LOCK_ERROR,
+    type: types.LOCK_ERROR,
     err,
   };
 }
