+++ client/src/containers/login.js
@@ -1,9 +1,9 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
-import Login from '../components/login';
+import LoginButton from './login_button';
 
 
-class App extends Component {
+class Login extends Component {
 
   static propTypes = {
     dispatch: PropTypes.func.isRequired,
@@ -16,7 +16,8 @@ class App extends Component {
 
     return (
       <div>
-        <Login
+        <h1>Stormbraining</h1>
+        <LoginButton
           isAuthenticated={isAuthenticated}
           errorMessage={errorMessage}
           dispatch={dispatch}
@@ -26,15 +27,12 @@ class App extends Component {
   }
 }
 
-function mapStateToProps(state) {
-
-  const { auth } = state;
+function mapStateToProps({ auth }) {
   const { isAuthenticated, errorMessage } = auth;
-
   return {
     isAuthenticated,
     errorMessage,
   };
 }
 
-export default connect(mapStateToProps)(App);
+export default connect(mapStateToProps)(Login);
